+++ package.json
@@ -11,6 +11,7 @@
     "@fortawesome/free-brands-svg-icons": "^5.14.0",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
+    "@material-ui/core": "^4.11.2",
     "@mdx-js/mdx": "^1.6.17",
     "@mdx-js/react": "^1.6.17",
     "emotion-theming": "^10.0.27",
