+++ src/content/about.mdx
@@ -1,4 +1,4 @@
-I'm a software engineer who lives and works in Seattle.
+I'm a software engineer who lives and works in the Seattle area.
 Growing up, I wanted to be a scientist like my Dad. In college, I pursued
 a dual major in Biochemistry and Computer Science so that I could work
 with data over their entire lifecycle.
